test(l1): move BAL input-hardening tests to root ethrex-test crate
Relocate the 0x-prefix serde tests and the build_validation_index
capacity-cap test out of inline #[cfg(test)] modules into
test/tests/common/bal_input_tests.rs, per repo convention that tests
live in the ethrex-test crate. Deserializers are exercised directly via
serde_json::Value (no serde-derive wrapper needed).
