+++ crates/common/serde_utils.rs
@@ -774,62 +774,3 @@ pub mod block_access_list {
         }
     }
 }
-
-#[cfg(test)]
-mod block_access_list_serde_tests {
-    use crate::types::block_access_list::BlockAccessList;
-    use ethrex_rlp::encode::RLPEncode;
-    use serde::Deserialize;
-
-    #[derive(Deserialize)]
-    struct WrapReq {
-        #[serde(with = "crate::serde_utils::block_access_list::rlp_str")]
-        bal: BlockAccessList,
-    }
-
-    #[derive(Deserialize)]
-    struct WrapOpt {
-        #[serde(with = "crate::serde_utils::block_access_list::rlp_str_opt")]
-        bal: Option<BlockAccessList>,
-    }
-
-    /// `0x` + RLP hex of an empty BAL (`0xc0`).
-    fn empty_bal_hex() -> String {
-        format!(
-            "0x{}",
-            hex::encode(BlockAccessList::from_accounts(vec![]).encode_to_vec())
-        )
-    }
-
-    #[test]
-    fn rlp_str_accepts_prefixed_and_rejects_unprefixed() {
-        let hex = empty_bal_hex();
-        let ok: WrapReq = serde_json::from_value(serde_json::json!({ "bal": hex })).unwrap();
-        assert_eq!(ok.bal.accounts().len(), 0);
-
-        // Same bytes without the mandatory 0x prefix must be rejected.
-        let unprefixed = hex.trim_start_matches("0x").to_string();
-        let err = serde_json::from_value::<WrapReq>(serde_json::json!({ "bal": unprefixed }));
-        assert!(err.is_err(), "unprefixed BAL hex must be rejected");
-    }
-
-    #[test]
-    fn rlp_str_opt_prefix_and_absence_handling() {
-        // "0x" encodes absence → None.
-        let none: WrapOpt = serde_json::from_value(serde_json::json!({ "bal": "0x" })).unwrap();
-        assert!(none.bal.is_none());
-
-        // JSON null → None.
-        let null: WrapOpt = serde_json::from_value(serde_json::json!({ "bal": null })).unwrap();
-        assert!(null.bal.is_none());
-
-        // A prefixed valid BAL → Some.
-        let some: WrapOpt =
-            serde_json::from_value(serde_json::json!({ "bal": empty_bal_hex() })).unwrap();
-        assert!(some.bal.is_some());
-
-        // Unprefixed (schema-invalid) → error, not silently accepted.
-        let err = serde_json::from_value::<WrapOpt>(serde_json::json!({ "bal": "c0" }));
-        assert!(err.is_err(), "unprefixed BAL hex must be rejected");
-    }
-}
