+++ crates/common/types/block_access_list.rs
@@ -1812,36 +1812,6 @@ mod decode_tests {
 }
 
 #[cfg(test)]
-mod build_index_tests {
-    use super::*;
-    use ethereum_types::Address;
-
-    /// The capacity hints in `build_validation_index` are capped against a
-    /// supplied BAL's (attacker-controlled) account count, but the cap is only a
-    /// pre-allocation hint: a block with more accounts than the cap must still be
-    /// indexed in full. Binds the `PREALLOC_CAP` min() against silent truncation.
-    #[test]
-    fn build_validation_index_indexes_more_accounts_than_prealloc_cap() {
-        let n = 8192usize + 5; // > PREALLOC_CAP
-        let accounts: Vec<AccountChanges> = (0..n)
-            .map(|i| {
-                let mut a = Address::zero();
-                a.0[16..20].copy_from_slice(&(i as u32).to_be_bytes());
-                AccountChanges::new(a)
-                    .with_balance_changes(vec![BalanceChange::new(1, U256::from(i as u64 + 1))])
-            })
-            .collect();
-        let bal = BlockAccessList::from_accounts(accounts);
-        let index = bal.build_validation_index();
-        assert_eq!(
-            index.addr_to_idx.len(),
-            n,
-            "every account must be indexed despite the capacity-hint cap"
-        );
-    }
-}
-
-#[cfg(test)]
 mod synthesize_tests {
     use super::*;
     use bytes::Bytes;
